[front] drop Poke message consumption breakdown
