+++ front-api/routes/poke/workspaces/[wId]/conversations/[cId]/index.ts
@@ -8,7 +8,6 @@ import { z } from "zod";
 
 import config from "./config";
 import consumption from "./consumption";
-import messageConsumption from "./messages/[mId]/consumption";
 import reinforcementTestCase from "./reinforcement_test_case";
 import render from "./render";
 
@@ -42,7 +41,6 @@ app.get(
 
 app.route("/config", config);
 app.route("/consumption", consumption);
-app.route("/messages/:mId/consumption", messageConsumption);
 app.route("/reinforcement_test_case", reinforcementTestCase);
 app.route("/render", render);
